40 feature 투표 기능

+++ src/main/java/com/example/academy/dto/vote/GetVoteInfoDTO.java
@@ -35,6 +35,6 @@ public class GetVoteInfoDTO {
     private String optionText;
     private String occupancyRate;
     private String votes;
-    private int rank;
+    private Long rank;
   }
 }
